double -> float again. We decided to keep the original formulation by now.

+++ include/util/realheap.h
@@ -11,7 +11,7 @@
 
 
 typedef struct _realheap {
-  double *cost;
+  float *cost;
   char *color;
   int *pixel;
   int *pos;
@@ -24,11 +24,11 @@ typedef struct _realheap {
 void SetRemovalPolicyRealHeap(RealHeap *H, char policy);
 char IsFullRealHeap(RealHeap *H);
 char IsEmptyRealHeap(RealHeap *H);
-RealHeap *CreateRealHeap(int n, double *cost);
+RealHeap *CreateRealHeap(int n, float *cost);
 void DestroyRealHeap(RealHeap **H);
 char InsertRealHeap(RealHeap *H, int pixel);
 char RemoveRealHeap(RealHeap *H, int *pixel);
-void UpdateRealHeap(RealHeap *H, int p, double value);
+void UpdateRealHeap(RealHeap *H, int p, float value);
 void GoUpRealHeap(RealHeap *H, int i);
 void GoDownRealHeap(RealHeap *H, int i);
 void ResetRealHeap(RealHeap *H);
